+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project 
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 4.4.1
 ### Added
 
 - Migration to [django-money](https://github.com/django-money/django-money), currency-aware
   representation of money values.
 - Added currency to all forms and views.
+- Switched default print template to match paper with full letterhead.
 
 ## 4.4.0
 ### Added